+++ www/go/core/install/install.sql
@@ -870,3 +870,24 @@ ALTER TABLE `core_email_template_attachment`
 
 ALTER TABLE `core_search` ADD INDEX(`keywords`);
 ALTER TABLE `core_change` ADD INDEX(`entityId`);
+
+
+
+CREATE TABLE `core_auth_allow_group` (
+  `id` int(11) NOT NULL,
+  `groupId` int(11) NOT NULL,
+  `ipPattern` varchar(50) COLLATE utf8mb4_unicode_ci NOT NULL COMMENT 'IP Address. Wildcards can be used where * matches anything and ? matches exactly one character'
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
+
+
+ALTER TABLE `core_auth_allow_group`
+  ADD PRIMARY KEY (`id`),
+  ADD KEY `groupId` (`groupId`);
+
+
+ALTER TABLE `core_auth_allow_group`
+  MODIFY `id` int(11) NOT NULL AUTO_INCREMENT;
+
+
+ALTER TABLE `core_auth_allow_group`
+  ADD CONSTRAINT `core_auth_allow_group_ibfk_1` FOREIGN KEY (`groupId`) REFERENCES `core_auth_allow_group` (`id`) ON DELETE CASCADE;