+++ www/go/core/views/extjs3/Module.js
@@ -24,7 +24,13 @@ go.Modules.register("core", 'core', {
 		'Search', 
 		'EntityFilter',
 		'SmtpAccount',
-		'EmailTemplate'
+		'EmailTemplate',
+		{
+			name: 'AllowGroup',
+			relations: {
+				group: {store: "Group", fk:'groupId'}
+			}
+		}
 	],
 
 	userSettingsPanels: [