+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Address book: Added filter for contacts being in a group or not
+
 19-12-2019 6.4.94
 - Core: Updated Brasilian Portugese language
 - Custom fields: There were 2 extra decimals for numbers in templates